pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oi-speaker"
version = "0.1.0"
requires-python = ">= 3.10"
dependencies = [
    "pyaudio",
    "sounddevice",
    "openwakeword @ git+https://github.com/dscripka/openWakeWord.git",
    "faster-whisper",
    "webrtcvad",
    "setuptools<71; sys_platform == 'darwin'",
    "anthropic",
    "piper-tts",
    "yt-dlp",
    "python-mpv",
    "speexdsp-ns; sys_platform == 'linux'",
    "fastapi[standard]",
    "tomli-w",
    "zeroconf>=0.131.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
training = [
    "scipy",
    "scikit-learn",
    "torch",
    "torchaudio",
    "onnx",
    "onnxscript",
    "soxr",
    "matplotlib",
]
cuda = [
    "nvidia-cublas-cu12",
    "nvidia-cudnn-cu12",
    "nvidia-cuda-runtime-cu12",
]

[project.scripts]
speaker = "speaker:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/speaker.py", "src/web.py", "src/static"]